Add guard for FileUtils directory removal to avoid error

+++ lib/steps/commit/scm_polling.rb
@@ -3,7 +3,7 @@ require 'fileutils'
 module Build
   module Commit
     class ScmPolling
-      ERROR_MESSAGE = 'Ensure a %s environment variable exists'
+      ERROR_MESSAGE = 'Ensure the %s environment variable exists'
 
       def initialize(store:)
         app_name = ENV['app_name'] || fail(ERROR_MESSAGE % ['app_name'])
@@ -16,7 +16,7 @@ module Build
         aws_azs = ENV['aws_azs'] || 'us-east-1c,us-east-1b'
         aws_keypair = ENV['aws_keypair'] || fail(ERROR_MESSAGE % ['aws_keypair'])
 
-        FileUtils.rm_r(working_directory)
+        FileUtils.rm_r(working_directory) if File.directory?(working_directory)
         git_cmd = "git clone --branch #{repository_branch} --depth 1 #{repository_url} #{working_directory}"
         system(git_cmd)
